# baseline/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/db/invoices": {
      "post": {
        "operationId": "InvoiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      }
    },
    "/api/db/invoices/pending-contracts": {
      "get": {
        "operationId": "InvoiceController_listPending",
        "parameters": [
          {
            "name": "tenantSchema",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campusId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      }
    },
    "/api/db/invoices/{id}": {
      "get": {
        "operationId": "InvoiceController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSchema",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invoice"
        ]
      }
    }
  },
  "info": {
    "title": "Edu Server API",
    "description": "教培 CRM 后端 API — Phase B.L3 contract source-of-truth",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "tenantSchema": {
            "type": "string",
            "description": "多租户 schema（TenantScopeGuard 校验）"
          },
          "invoiceId": {
            "type": "string",
            "description": "32-char ULID（前端生成；与 contracts/students 一致）",
            "minLength": 32,
            "maxLength": 32
          },
          "contractId": {
            "type": "string",
            "description": "OOUX 父对象 contract.id（32-char ULID）",
            "minLength": 32,
            "maxLength": 32
          },
          "titleType": {
            "type": "string",
            "description": "抬头类型",
            "enum": [
              "个人",
              "企业"
            ]
          },
          "invoiceTitle": {
            "type": "string",
            "description": "开票抬头（≤ 80，msgSecCheck 过 wx.security）",
            "maxLength": 80
          },
          "taxId": {
            "type": "string",
            "description": "企业必填；18 位统一信用代码",
            "maxLength": 32
          },
          "receiveEmail": {
            "type": "string",
            "description": "接收邮箱（必填）"
          },
          "receivePhone": {
            "type": "string",
            "description": "接收手机号（可选；走 hash + encrypted 三写）"
          },
          "remark": {
            "type": "string",
            "description": "备注（≤ 200，msgSecCheck）",
            "maxLength": 200
          }
        },
        "required": [
          "tenantSchema",
          "invoiceId",
          "contractId",
          "titleType",
          "invoiceTitle",
          "receiveEmail"
        ]
      }
    }
  }
}